Dritbra triks å lansere nye nettsider midt i Desember!

+++ vindjulekalender.sh
@@ -2,7 +2,7 @@
 
 TODAY=$(date +%-d)
 FULLDATE=$(date +%d.%m.%Y)
-URL="https://www.vindil.no/wp-content/uploads/2025/12/Trekning-$TODAY.12.2025-1080x675.jpg"
+URL="https://vindil.no/storage/media/185/conversions/Trekning-$FULLDATE-main.jpg"
 TMPFILE="/tmp/vind-${TODAY}.jpg"
 MYNUMBER=$1
 IWON=0
@@ -17,7 +17,7 @@ function box() { local t="$1xxxx";local c="${2:-=}"; echo "${t//?/$c}"; echo "$c
 box "Vinnertall i Vind ILs julekalender $FULLDATE "
 wget -q "$URL" -O $TMPFILE
 if [[ -s $TMPFILE ]]; then
-  WINNERS=$(tesseract $TMPFILE - | grep -oE '^[0-9]{3,4}')
+  WINNERS=$(tesseract $TMPFILE - quiet | grep -oE '^[0-9]{2,4}:' | tr -d ':')
   if [[ -n "$MYNUMBER" ]]; then
     echo "Ditt registrerte loddnummer er: $MYNUMBER"
     $(echo "$WINNERS" | grep -q $MYNUMBER) && IWON=1
@@ -26,9 +26,11 @@ else
   WINNERS='...ikke trukket enda...'
 fi
 
- echo "Dagens tall: ${WINNERS//$'\n'/ }"
+echo "Dagens tall: ${WINNERS//$'\n'/ }"
 if [[ $IWON -eq 1 ]]; then
   echo "Du vant!"
 fi
 
-rm $TMPFILE
+if [ -f $TMPFILE ]; then
+  rm $TMPFILE
+fi
